+++ lib/tmuxinator/config.rb
@@ -27,6 +27,10 @@ module Tmuxinator
         File.directory?(home)
       end
 
+      def options_file
+        directory + ".yml"
+      end
+
       # ~/.config/tmuxinator unless $XDG_CONFIG_HOME has been configured to use
       # a custom value. (e.g. if $XDG_CONFIG_HOME is set to ~/my-config, the
       # return value will be ~/my-config/tmuxinator)
@@ -61,6 +65,12 @@ module Tmuxinator
 
       def default
         "#{directory}/default.yml"
+      end
+
+      def options
+        YAML.safe_load(File.read(options_file), aliases: true) || {}
+      rescue SyntaxError, StandardError
+        {}
       end
 
       def default?